fix pcap-usb-linux.c size_t

+++ pcap-usb-linux-common.c
@@ -56,7 +56,7 @@ fix_linux_usb_mmapped_length(struct pcap_pkthdr *pkth, const u_char *bp)
 	                 (hdr->ndesc * sizeof (usb_isodesc)) + hdr->urb_len) {
 		usb_isodesc *descs;
 		u_int pre_truncation_data_len;
-		unsigned long pre_truncation_len;
+		size_t pre_truncation_len;
 
 		descs = (usb_isodesc *) (bp + sizeof(pcap_usb_header_mmapped));
